fix(new-agent): address post-review issues in Create-with-AI flow
- Derive FOCUS_KEYS/TRAIT_KEYS from label maps to eliminate slug
  duplication across three sites
- Extract serializeFormValues to agent-setup-utils.ts; form file
  now under 200-line limit
- Guard heading prepend: add # Name only when engine omits it,
  preventing double-heading without losing it on inconsistent responses
- Add creating state + spinner to AiReviewStep; prevents double-submit
  on slow network
- Restore eslint-disable comment on useEffect in ai-review-step
- Fix broken Portuguese: Quão, solicitação, Instruções, Analítico, etc.
- Add comment explaining personal-assistant pin offset in store-step
