+++ app/src/locales/pt/shell.json
@@ -190,32 +190,32 @@
       "focusRequired": "Escolha uma área de foco para continuar",
       "traitsLabel": "Escolha até 3 características",
       "traitsHint": "{{selected}}/{{max}} selecionadas",
-      "verbosityLabel": "Quao detalhadas devem ser as respostas?",
+      "verbosityLabel": "Quão detalhadas devem ser as respostas?",
       "verbosityMin": "Breve",
       "verbosityMax": "Exaustivo",
       "askFirstLabel": "Deve fazer perguntas antes de responder?",
-      "askFirstHint": "O agente vai esclarecer sua solicitacao antes de comecar",
-      "extraLabel": "Instrucoes adicionais",
+      "askFirstHint": "O agente vai esclarecer sua solicitação antes de começar",
+      "extraLabel": "Instruções adicionais",
       "extraOptional": "opcional",
-      "extraPlaceholder": "Outras preferencias, regras ou contexto para este agente..."
+      "extraPlaceholder": "Outras preferências, regras ou contexto para este agente..."
     },
     "focus": {
       "personalProductivity": "Produtividade pessoal",
-      "workBusiness": "Trabalho e negocios",
+      "workBusiness": "Trabalho e negócios",
       "researchLearning": "Pesquisa e aprendizado",
-      "writingEditing": "Escrita e edicao",
-      "codingDev": "Programacao e desenvolvimento",
+      "writingEditing": "Escrita e edição",
+      "codingDev": "Programação e desenvolvimento",
       "customerSupport": "Suporte ao cliente",
-      "dataAnalysis": "Dados e analise",
+      "dataAnalysis": "Dados e análise",
       "creativeDesign": "Criatividade e design"
     },
     "traits": {
       "concise": "Conciso",
       "thorough": "Detalhado",
-      "friendly": "Amigavel",
+      "friendly": "Amigável",
       "professional": "Profissional",
       "creative": "Criativo",
-      "analytical": "Analitico",
+      "analytical": "Analítico",
       "proactive": "Proativo",
       "cautious": "Cauteloso",
       "structured": "Estruturado",
